+++ src/main/java/org/opensearch/knn/indices/ModelUtil.java
@@ -12,7 +12,7 @@
 package org.opensearch.knn.indices;
 
 import lombok.experimental.UtilityClass;
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 
 import java.util.Locale;